Cancel any async work started on the DocumentPaginator when it changes
GetPageAsync is called from MeasureOverride when _useAsynchronous is true, but this work isn't canceled when the DocumentPaginator changes. In the case where you are using the paginator returned from an XpsDocument's GetFixedDocumentSequence method, if you update the DocumentPaginator on this control, then delete the backing file used by the old XpsDocument's document paginator, it can result in an exception while continuing to load the page since the async work was not canceled.